requirements.txt
discord.py
requests
discord-py-interactions